refactor upgrades with feedback
