+++ anvil-python/anvil/commands/haproxy.py
@@ -30,7 +30,6 @@ from sunbeam.jobs.steps import (
 
 from anvil.jobs.manifest import Manifest
 from anvil.jobs.steps import RemoveMachineUnitStep
-from anvil.utils import UpgradeCharm
 
 LOG = logging.getLogger(__name__)
 
@@ -185,28 +184,6 @@ class RemoveHAProxyUnitStep(RemoveMachineUnitStep):
 
     def get_unit_timeout(self) -> int:
         return HAPROXY_UNIT_TIMEOUT
-
-
-class UpgradeHAProxyUnitCharms(UpgradeCharm):
-    def __init__(
-        self,
-        client: Client,
-        jhelper: JujuHelper,
-        manifest: Manifest,
-        model: str,
-    ):
-        super().__init__(
-            "Upgrade HAProxy unit charms",
-            "Upgrading HAProxy unit charms.",
-            client,
-            jhelper,
-            manifest,
-            model,
-            ["haproxy", "keepalived"],
-            "haproxy-plan",
-            HAPROXY_CONFIG_KEY,
-            HAPROXY_UNIT_TIMEOUT,
-        )
 
 
 def haproxy_install_steps(
