+++ anvil-python/anvil/commands/maas_region.py
@@ -26,7 +26,6 @@ from sunbeam.jobs.steps import (
 
 from anvil.jobs.manifest import Manifest
 from anvil.jobs.steps import RemoveMachineUnitStep
-from anvil.utils import UpgradeCharm
 
 APPLICATION = "maas-region"
 CONFIG_KEY = "TerraformVarsMaasregionPlan"
@@ -118,28 +117,6 @@ class RemoveMAASRegionUnitStep(RemoveMachineUnitStep):
 
     def get_unit_timeout(self) -> int:
         return MAASREGION_UNIT_TIMEOUT
-
-
-class UpgradeMAASRegionUnitCharms(UpgradeCharm):
-    def __init__(
-        self,
-        client: Client,
-        jhelper: JujuHelper,
-        manifest: Manifest,
-        model: str,
-    ):
-        super().__init__(
-            "Upgrade MAAS Region unit charms",
-            "Upgrading MAAS Region unit charms.",
-            client,
-            jhelper,
-            manifest,
-            model,
-            ["maas-region", "pgbouncer"],
-            "maas-region-plan",
-            CONFIG_KEY,
-            MAASREGION_UNIT_TIMEOUT,
-        )
 
 
 def maas_region_install_steps(
