+++ anvil-python/anvil/commands/postgresql.py
@@ -29,7 +29,6 @@ from sunbeam.jobs.steps import (
 
 from anvil.jobs.manifest import Manifest
 from anvil.jobs.steps import RemoveMachineUnitStep
-from anvil.utils import UpgradeCharm
 
 LOG = logging.getLogger(__name__)
 APPLICATION = "postgresql"
@@ -278,25 +277,3 @@ class RemovePostgreSQLUnitStep(RemoveMachineUnitStep):
 
     def get_unit_timeout(self) -> int:
         return POSTGRESQL_UNIT_TIMEOUT
-
-
-class UpgradePostgreSQLUnitCharms(UpgradeCharm):
-    def __init__(
-        self,
-        client: Client,
-        jhelper: JujuHelper,
-        manifest: Manifest,
-        model: str,
-    ):
-        super().__init__(
-            "Upgrade PostgreSQL unit charms",
-            "Upgrading PostgreSQL unit charms.",
-            client,
-            jhelper,
-            manifest,
-            model,
-            ["postgresql"],
-            "postgresql-plan",
-            CONFIG_KEY,
-            POSTGRESQL_UNIT_TIMEOUT,
-        )
